# packages/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona API",
    "summary": "Hosted service for building and running typed-memory AI personas.",
    "version": "0.8.0"
  },
  "paths": {
...
    "/v1/mcp-servers": {
      "get": {
        "tags": ["mcp"],
        "summary": "List Mcp Servers",
        "description": "List the caller's BYO MCP servers (credential redacted).",
        "operationId": "list_mcp_servers_v1_mcp_servers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MCPServerDetail"
                  },
                  "type": "array",
                  "title": "Response List Mcp Servers V1 Mcp Servers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["mcp"],
        "summary": "Create Mcp Server",
        "description": "Add a bring-your-own MCP server (SSRF-validated; credential encrypted).",
        "operationId": "create_mcp_server_v1_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMCPServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-servers/{server_id}": {
      "get": {
        "tags": ["mcp"],
        "summary": "Get Mcp Server",
        "description": "Get one BYO MCP server (RLS-scoped \u2192 404).",
        "operationId": "get_mcp_server_v1_mcp_servers__server_id__get",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["mcp"],
        "summary": "Update Mcp Server",
        "description": "Patch a BYO MCP server (re-validates a new URL; re-encrypts credentials).",
        "operationId": "update_mcp_server_v1_mcp_servers__server_id__patch",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMCPServerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["mcp"],
        "summary": "Delete Mcp Server",
        "description": "Delete a BYO MCP server (assignments cascade; RLS-scoped \u2192 404).",
        "operationId": "delete_mcp_server_v1_mcp_servers__server_id__delete",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-servers/{server_id}/test": {
      "post": {
        "tags": ["mcp"],
        "summary": "Check Mcp Server Connection",
        "description": "Test-connect to the server (SSRF-pinned) and discover its tools (D-30-5).",
        "operationId": "check_mcp_server_connection_v1_mcp_servers__server_id__test_post",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServerTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/{persona_id}/mcp-servers": {
      "get": {
        "tags": ["mcp"],
        "summary": "List Persona Mcp Servers",
        "description": "List the BYO MCP servers assigned to a persona (RLS-scoped).",
        "operationId": "list_persona_mcp_servers_v1_personas__persona_id__mcp_servers_get",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MCPServerDetail"
                  },
                  "title": "Response List Persona Mcp Servers V1 Personas  Persona Id  Mcp Servers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/{persona_id}/mcp-servers/{server_id}": {
      "put": {
        "tags": ["mcp"],
        "summary": "Assign Mcp Server",
        "description": "Assign a BYO MCP server to a persona (D-30-6; idempotent; RLS both ends).",
        "operationId": "assign_mcp_server_v1_personas__persona_id__mcp_servers__server_id__put",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          },
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["mcp"],
        "summary": "Unassign Mcp Server",
        "description": "Remove a persona\u2194server assignment (idempotent; RLS-scoped).",
        "operationId": "unassign_mcp_server_v1_personas__persona_id__mcp_servers__server_id__delete",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          },
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/personas/{persona_id}/adopted-apps": {
      "post": {
        "tags": ["mcp"],
        "summary": "Adopt Catalog App",
        "description": "Self-adopt a catalog app for a persona (Spec N4, B2-\u2462).\n\nOwner-scoped (the persona must be the caller's \u2192 404) and vetted (N4-D-6 \u2192 403),\nboth BEFORE any write. The connection url/auth are derived from the catalog entry\n(N4-D-10); the caller supplies only ``credential`` (a ``repr=False`` field, encrypted\nat rest, never returned/logged). The audit records name + provenance only.",
        "operationId": "adopt_catalog_app_v1_personas__persona_id__adopted_apps_post",
        "parameters": [
          {
            "name": "persona_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Persona Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptCatalogAppRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdoptCatalogAppRequest": {
        "properties": {
          "catalog_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Catalog Name"
          },
          "credential": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["catalog_name"],
        "title": "AdoptCatalogAppRequest",
        "description": "Self-adopt a catalog app for a persona (Spec N4, the B2-\u2462 setup-form target).\n\nThe connection ``url`` and ``auth_method`` are derived from the catalog entry\nserver-side (N4-D-10 \u2014 the catalog is the trust anchor for *where* it connects); the\ncaller supplies ONLY ``credential`` (when the app declares a secret). ``credential`` is\n``repr=False`` (redacted in logs), encrypted at rest via the store, and NEVER returned."
      },
...
      "CreateMCPServerRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          },
          "auth_method": {
            "type": "string",
            "enum": ["none", "bearer"],
            "title": "Auth Method",
            "default": "none"
          },
          "credential": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["name", "url"],
        "title": "CreateMCPServerRequest",
        "description": "Add a bring-your-own MCP server (spec 30, D-30-3/4).\n\n``url`` is SSRF-validated (https-only, public target) at the route AND on\nevery live connect. ``credential`` (a bearer token for ``auth_method =\n\"bearer\"``) is encrypted at rest (T07) and NEVER returned or logged; it is\nrequired when ``auth_method`` is not ``\"none\"``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MCPServerDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "auth_method": {
            "type": "string",
            "title": "Auth Method"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "has_credential": {
            "type": "boolean",
            "title": "Has Credential"
          },
          "discovered_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovered Tools"
          },
          "catalog_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "url",
          "auth_method",
          "enabled",
          "has_credential",
          "created_at",
          "updated_at"
        ],
        "title": "MCPServerDetail",
        "description": "A bring-your-own MCP server as returned to its owner (spec 30, D-30-3).\n\nThe credential is NEVER included \u2014 only ``has_credential`` (whether one is\nstored). ``discovered_tools`` is the cached eager-discovery result (D-30-5),\n``None`` until a successful test-connection."
      },
      "MCPServerTestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["ok"],
        "title": "MCPServerTestResult",
        "description": "Outcome of a BYO-MCP test-connection (spec 30, D-30-5).\n\n``ok`` true \u2192 ``tools`` lists the discovered tool names (cached on the row).\n``ok`` false \u2192 ``error`` is a short, non-sensitive reason category."
      },
...
      "UpdateMCPServerRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "auth_method": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["none", "bearer"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Method"
          },
          "credential": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateMCPServerRequest",
        "description": "Patch a BYO MCP server (spec 30). All fields optional; omitted = unchanged.\n\nSetting ``credential`` replaces the stored secret (re-encrypted); to clear a\ncredential, set ``auth_method = \"none\"``. ``enabled`` toggles the server\nwithout deleting it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}